backend/requirements.txt
Django>=5.1.6,<6.0
djangorestframework>=3.14.0
django-cors-headers>=3.13.0
djoser>=2.1.0
djangorestframework-simplejwt>=5.2.2
celery>=5.2.3
psycopg2-binary>=2.9.3
gunicorn>=20.1.0
redis>=3.5.3
openpyxl
pandas>=1.0.0